[Robot] Read binary packed data from TagTracker

+++ Robot/src/main/java/com/swrobotics/robot/subsystems/tagtracker/io/TagTrackerCameraIO.java
@@ -9,6 +9,6 @@ public interface TagTrackerCameraIO {
 
     final class Inputs {
         public long[] timestamps;
-        public double[][] framePackedData;
+        public byte[][] framePackedData;
     }
 }
